k8s-client: EventsApiClientFunctionalTest - finalisation
