+++ libs/k8s-client/tests/ApiClient/EventsApiClientFunctionalTest.php
@@ -5,97 +5,50 @@ declare(strict_types=1);
 namespace Keboola\K8sClient\Tests\ApiClient;
 
 use Keboola\K8sClient\ApiClient\EventsApiClient;
-use Keboola\K8sClient\ClientFacadeFactory\GenericClientFacadeFactory;
-use Keboola\K8sClient\KubernetesApiClientFacade;
-use Keboola\K8sClient\RetryProxyFactory;
+use Kubernetes\API\Event as EventsApi;
 use Kubernetes\Model\Io\K8s\Api\Core\V1\Event;
-use Kubernetes\Model\Io\K8s\Api\Core\V1\Pod;
-use Kubernetes\Model\Io\K8s\Apimachinery\Pkg\Apis\Meta\V1\DeleteOptions;
-use Monolog\Logger;
 use PHPUnit\Framework\TestCase;
 
 class EventsApiClientFunctionalTest extends TestCase
 {
-    private static KubernetesApiClientFacade $apiClient;
-    private static EventsApiClient $eventsApiClient;
+    /**
+     * @template-use BaseNamespaceApiClientTestCase<EventsApi, EventsApiClient>
+     */
+    use BaseNamespaceApiClientTestCase;
 
-    public static function setUpBeforeClass(): void
+    public function setUp(): void
     {
-        parent::setUpBeforeClass();
-
-        $logger = new Logger('test');
-
-        $clientFacade = (new GenericClientFacadeFactory(
-            (new RetryProxyFactory($logger))->createRetryProxy(),
-            $logger
-        ))->createClusterClient(
-            (string) getenv('K8S_HOST'),
-            (string) getenv('K8S_TOKEN'),
-            (string) getenv('K8S_CA_CERT_PATH'),
-            (string) getenv('K8S_NAMESPACE'),
-        );
-
-        self::$apiClient = $clientFacade;
-        self::$eventsApiClient = $clientFacade->events();
-
-        // generate a few events to test below
-        self::cleanupCluster();
-
-        $clientFacade->pods()->create(new Pod([
-            'metadata' => [
-                'name' => 'test-pod',
-                'labels' => [
-                    'app' => 'EventsApiClientFunctionalTest',
-                ],
-            ],
-            'spec' => [
-                'containers' => [
-                    [
-                        'name' => 'nginx',
-                        'image' => 'nginx',
-                    ],
-                ],
-            ],
-        ]));
-    }
-
-    public static function tearDownAfterClass(): void
-    {
-        parent::tearDownAfterClass();
-        self::cleanupCluster();
-    }
-
-    private static function cleanupCluster(): void
-    {
-        self::$apiClient->deleteAllMatching(
-            new DeleteOptions(['gracePeriodSeconds' => 0]),
-            ['labelSelector' => 'app=EventsApiClientFunctionalTest']
+        parent::setUp();
+        $this->setUpBaseNamespaceApiClientTest(
+            EventsApi::class,
+            EventsApiClient::class,
         );
     }
 
-    public function testList(): void
+    protected function createResource(array $metadata): Event
     {
-        $result = self::$eventsApiClient->list();
-        self::assertGreaterThan(0, count($result->items));
-    }
-
-    public function testListWithQuery(): void
-    {
-        //@TODO move logic to BaseNamespaceApiClientTestCase
-        $result = self::$eventsApiClient->list([
-            'fieldSelector' => 'involvedObject.name=test-pod',
+        return new Event([
+            'metadata' => $metadata,
+            'type' => 'Normal',
+            'reason' => 'Testing',
+            'message' => 'EventsErikApiClientFunctionalTest event',
+            'involvedObject' => [
+                'kind' => 'ConfigMap',
+                'name' => 'kube-root-ca.crt',
+                'namespace' => (string) getenv('K8S_NAMESPACE'),
+            ],
         ]);
-
-        self::assertGreaterThan(0, count($result->items));
-        foreach ($result->items as $event) {
-            self::assertInstanceOf(Event::class, $event);
-            self::assertSame('test-pod', $event->involvedObject->name);
-        }
     }
 
     public function testListForAllNamespaces(): void
     {
-        $result = self::$eventsApiClient->listForAllNamespaces();
+        $result = $this->apiClient->listForAllNamespaces();
         self::assertGreaterThan(0, count($result->items));
+        $namespaces = [];
+        foreach ($result->items as $event) {
+            $namespaces[$event->metadata->namespace] = $event->metadata->namespace;
+        }
+
+        self::assertGreaterThan(1, $namespaces);
     }
 }
